LPS-55928 move dummy impls
