+++ portal-service/src/com/liferay/portal/kernel/search/dummy/DummyIndexSearcher.java
@@ -12,8 +12,16 @@
  * details.
  */
 
-package com.liferay.portal.kernel.search;
+package com.liferay.portal.kernel.search.dummy;
 
+import com.liferay.portal.kernel.search.Document;
+import com.liferay.portal.kernel.search.Hits;
+import com.liferay.portal.kernel.search.HitsImpl;
+import com.liferay.portal.kernel.search.IndexSearcher;
+import com.liferay.portal.kernel.search.Query;
+import com.liferay.portal.kernel.search.SearchContext;
+import com.liferay.portal.kernel.search.Sort;
+import com.liferay.portal.kernel.search.StringQueryImpl;
 import com.liferay.portal.kernel.util.StringPool;
 
 import java.util.Collections;
