+++ portal-service/src/com/liferay/portal/kernel/search/dummy/DummyIndexWriter.java
@@ -12,7 +12,11 @@
  * details.
  */
 
-package com.liferay.portal.kernel.search;
+package com.liferay.portal.kernel.search.dummy;
+
+import com.liferay.portal.kernel.search.Document;
+import com.liferay.portal.kernel.search.IndexWriter;
+import com.liferay.portal.kernel.search.SearchContext;
 
 import java.util.Collection;
